justecommerce/requirements.txt
asgiref==3.6.0
cachetools==5.3.1
chardet==5.1.0
colorama==0.4.6
distlib==0.3.6
Django==4.2.1
django-appconf==1.0.5
django-image-cropping==0.3.7
django-imagekit==4.1.0
django-mathfilters==1.0.0
django-resized==1.0.2
easy-thumbnails==2.8.5
filelock==3.12.2
numpy==1.25.1
packaging==23.1
pandas==2.0.3
pilkit==2.0
Pillow==9.5.0
platformdirs==3.8.1
pluggy==1.2.0
psycopg==3.1.9
psycopg2==2.9.6
pyproject-api==1.5.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tox==4.6.4
typing_extensions==4.6.0
tzdata==2023.3
virtualenv==20.23.1
virtualenvwrapper-win==1.2.7
